pyproject.toml
[project]
name = "docker-tui"
version = "0.1.0"
authors = [
    { name = "Jason Lo", email = "lcmjlo@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "textual>=8.0.0",
    "docker>=7.0.0",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "."

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[project.scripts]
docker_tui = "docker_tui.app:main"
